Continued on update pheromone.

+++ Basic/Constants.cs
@@ -5,6 +5,6 @@
         // Initial value of pheromone on graph. C
         public static readonly double MinimalValueOfPheromone = 0.00001D;
         //
-        public static readonly decimal MinimalVelueOfPheromoneToSet = 0.0000001M;
+        public static readonly double MinimalVelueOfPheromoneToSet = 0.0000001D;
     }
 }
